+++ CHANGES.md
@@ -9,6 +9,7 @@ Updated:
 - SillyTavern
 - WhisperSpeech web UI
 - F5-TTS
+- ComfyUI
 
 Added:
 - HeartMuLa
@@ -18,3 +19,10 @@ Removed:
 - Matcha-TTS
 - Dia
 - KaniTTS
+
+ComfyUI:
+- Manager and ComfyUI-GGUF are installed by default
+- Qwen-Image-Edit and Qwen-Image-Edit-2509 replaced by Qwen-Image-Edit-2511
+- Qwen-Image replaced by  Qwen-Image-Edit-2512
+- Removed Wan2.2-TI2V-5B
+- Added LTX-2-GGUF